% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accorpa_AAMMGG.R
\name{accorpa_AAMMGG}
\alias{accorpa_AAMMGG}
\title{Accorpa data Anno,Mese,Giorno in AnnoMeseGiorno}
\usage{
accorpa_AAMMGG(aa, mm, gg)
}
\arguments{
\item{aa}{Rappresenta l'anno}

\item{mm}{Rappresenta il mese}

\item{gg}{Rappresenta il giorno}
}
\value{
I tre campi accorpati
}
\description{
Nelle basi dati tipo i GEUTY, si trovano le date divise in tre campi diversi
come ad esempio DATAPERF_AA, DATAPERF_MM, DATAPERF_GG
La funzione serve per unirli in un unico campo.
Viene previsto anche il caso in cui anno, mese e giorno siano nulli
}
\examples{
data <- accorpa_AAMMGG(2008,1,19);
}
