% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connessione_as400.R
\name{connessione_as400}
\alias{connessione_as400}
\title{Connessione AS400}
\usage{
connessione_as400(configurazione)
}
\arguments{
\item{configurazione}{che deve avere userID, pwd, dbAS400}
}
\value{
Ritorna la connessione
}
\description{
Questa funzione serve per connettersi ad AS400
}
\examples{
con <- connessione_as400(configurazione);
}
