% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/definisci_data.R
\name{definisci_data}
\alias{definisci_data}
\title{Definisce date}
\usage{
definisci_data(quando = "Oggi", dt = Sys.Date())
}
\arguments{
\item{quando}{è il nome del tempo e può essere InizioAnno, FineMesePrecedente, Oggi (default), Ultimi3Mesi, Ultimi12Mesi}

\item{dt}{è una data di partenza. Di default Sys.Date()}
}
\value{
Ritorna la data desiderata
}
\description{
Questa funzione serve per definire date
}
\examples{
data <- definisci_data(quando="Ultimi3Mesi",dt="2022-12-06");
}
