% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estrai_tabella_geuty400f.R
\name{estrai_tabella_geuty400f}
\alias{estrai_tabella_geuty400f}
\title{Estrae GEUTY}
\usage{
estrai_tabella_geuty400f(connessione)
}
\arguments{
\item{connessione}{è la connessione}
}
\value{
Ritorna geuty
}
\description{
Questa funzione serve per estrarre geuty
}
\examples{
g400 <- estrai_tabella_geuty400f(connessione);
}
