% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estrai_tabella_geuty420f.R
\name{estrai_tabella_geuty420f}
\alias{estrai_tabella_geuty420f}
\title{Estrae GEUTY420F con gli importi erogati ed ergoabili}
\usage{
estrai_tabella_geuty420f(connessione)
}
\arguments{
\item{connessione}{è la connessione ad AS400 di solito realizzata con la funzione connessione_as400 del pacchetto.}
}
\value{
Ritorna geuty420f
}
\description{
Questa funzione permette di estrarre la tabella GEUTY420F dal CRM aziendale su AS400
}
\details{
La elaborazione dura almeno un paio di minuti perchè deve estrarre 
centinaia di migliaia di records.

Al termine viene estratto un dataframe sottoforma di tibble.

I campi numerici sono stati modificati correttamente in numero
}
\examples{
g420 <- estrai_tabella_geuty420f(connessione);
}
