% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leggi_configurazione.R
\name{leggi_configurazione}
\alias{leggi_configurazione}
\title{Leggi configurazione in formato YAML}
\usage{
leggi_configurazione(file_config)
}
\arguments{
\item{file_config}{è il nome del file da leggere}
}
\value{
Ritorna un elenco con le configurazioni da utilizzare
}
\description{
Questa funzione serve per leggere un file di configurazione YAML
}
\examples{
cnf <- leggi_configurazione("config.yaml");
}
