% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riprogramma_testo.R
\name{riprogramma_testo}
\alias{riprogramma_testo}
\title{Riprogramma il testo}
\usage{
riprogramma_testo(elemento)
}
\arguments{
\item{elemento}{è il nome dell'elemento da riprogrammare}
}
\value{
Ritorna un elemento riprogrammato
}
\description{
Questa funzione serve riprogrammare il testo
}
\examples{
elem <- riprogramma_testo(elemento);
}
