% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estrai_tabella_geuty400f.R
\name{estrai_tabella_geuty400f}
\alias{estrai_tabella_geuty400f}
\title{Estrae GEUTY}
\usage{
estrai_tabella_geuty400f(connessione)
}
\arguments{
\item{connessione}{è la connessione ad AS400 di solito realizzata con la funzione connessione_as400 del pacchetto.}
}
\value{
Ritorna geuty
}
\description{
Questa funzione permette di estrarre la tabella GEUTY400F dal CRM aziendale su AS400
}
\details{
La elaborazione dura almeno un paio di minuti perchè deve estrarre 
centinaia di migliaia di records.

Al termine viene estratto un dataframe sottoforma di tibble.

I campi numerici sono stati modificati correttamente in numero
}
\examples{
g400 <- estrai_tabella_geuty400f(connessione);
}
