% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estrai_tabella_isal.R
\name{estrai_tabella_isal}
\alias{estrai_tabella_isal}
\title{Estrai iSal}
\usage{
estrai_tabella_isal(connessione)
}
\arguments{
\item{connessione}{è connessione}
}
\value{
Ritorna iSal
}
\description{
Questa funzione serve per estrarre iSal
}
\examples{
isal <- estrai_tabella_isal(connessione);
}
