% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estrai_tabella_restfopecq.R
\name{estrai_tabella_restfopecq}
\alias{estrai_tabella_restfopecq}
\title{Estrai restfope}
\usage{
estrai_tabella_restfopecq(connessione)
}
\arguments{
\item{connessione}{è connessione}
}
\value{
Ritorna restfope
}
\description{
Questa funzione serve per estrarre restfope
}
\examples{
rst <- estrai_tabella_restfopecq(connessione);
}
