% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estrai_tabella_utcre001w2.R
\name{estrai_tabella_utcre001w2}
\alias{estrai_tabella_utcre001w2}
\title{Estrai utcre001w2}
\usage{
estrai_tabella_utcre001w2(connessione)
}
\arguments{
\item{connessione}{è connessione}
}
\value{
Ritorna utcre001w2
}
\description{
Questa funzione serve per estrarre utcre001w2
}
\examples{
ut <- estrai_tabella_utcre001w2(connessione);
}
